fixed scripts

+++ util/make_img.sh
@@ -1,5 +1,5 @@
 #!/bin/bash -e
-# set -o pipefail
+set -o pipefail
 
 if [[ -z $1 ]]; then
 	echo "Usage: $0 <output>"
@@ -18,9 +18,7 @@ set -x
 
 unmount-filesystems() {
 	sudo umount mnt || true
-	sudo umount root/var/local
-	sudo umount root/boot/efi
-	sudo umount root/boot
+	sudo umount root/boot/efi || true
 	sudo umount -R root/dev
 	sudo umount -R root/sys
 	sudo umount root/proc
@@ -50,56 +48,56 @@ tmp=$(mktemp -dp ~/.cache)
 
 cd "$tmp"
 
-# Bootstrap Debian 12.
-sudo debootstrap trixie root
+# Bootstrap Debian 13.
+# sudo debootstrap trixie root
+	sudo cp -a ~/Downloads/trixie/ root/
 
 # Create image.
-fallocate -l8G atlas.img
+if [[ $ATLAS_FS = ext ]]; then
+	image_size=3G
+else
+	image_size=1G
+fi
+fallocate -l$image_size atlas.img
 
 # Create partition table.
 sgdisk -o atlas.img
 
 # Create partitions.
 sgdisk -n 0:0:+64MiB -t 0:ef00 -c 0:esp atlas.img
-sgdisk -n 0:0:+128MiB -t 0:8300 -c 0:boot atlas.img
-sgdisk -n 0:0:+128MiB -t 0:8300 -c 0:local atlas.img
 
 # Create loopback.
 loopback=$(sudo losetup -fP --show atlas.img)
 
 # Create filesystems.
 sudo mkfs.fat "${loopback}p1"
-sudo mkfs.ext4 -F "${loopback}p2"
-sudo mkfs.ext4 -F "${loopback}p3"
 
 # Mount filesystems.
 sudo mount -t proc proc root/proc
 sudo mount --rbind --make-rslave /sys root/sys
 sudo mount --rbind --make-rslave /dev root/dev
-sudo mount "${loopback}p2" root/boot
 sudo mkdir -p root/boot/efi
 sudo mount "${loopback}p1" root/boot/efi
-sudo mount "${loopback}p3" root/var/local
 
 # Copy source code.
-sudo rsync -av --exclude={.dart_tool,.git,build,node_modules,target} "$src/" "root/root/atlas/"
+sudo rsync -a --exclude={.dart_tool,.git,build,node_modules,target} "$src/" root/root/atlas/
 
 # Install Atlas.
 sudo chroot root bash -e << EOF
-ATLAS_DEBUG=$ATLAS_DEBUG BOOTSTRAP=1 ~/atlas/util/bootstrap.sh
+ATLAS_DEBUG=$ATLAS_DEBUG ATLAS_FS=$ALTAS_FS BOOTSTRAP=1 ~/atlas/util/bootstrap.sh
 EOF
 
 unmount-filesystems
 
 # Add root partition.
-if [[ $ATLAS_DEBUG = 1 ]]; then
+if [[ $ATLAS_FS = ext ]]; then
 	# Make writable ext4 rootfs.
-	sgdisk -n 0:0:+6GiB -t 0:8304 -c 0:root atlas.img
+	sgdisk -n 0:0:+2GiB -t 0:8304 -c 0:root atlas.img
 	sudo partprobe "$loopback"
-	sudo mkfs.ext4 -F "${loopback}p4"
+	sudo mkfs.ext4 -F "${loopback}p2"
 	mkdir -p mnt
-	sudo mount "${loopback}p4" mnt
-	sudo rsync -ra root/ mnt/
+	sudo mount "${loopback}p2" mnt
+	sudo cp -a root/ mnt/
 	sudo umount mnt
 else
 	# Make readonly squashfs root.
@@ -108,13 +106,21 @@ else
 	sqfs_size_kib=$(((sqfs_size_b+1023)/1024))
 	sgdisk -n 0:0:+${sqfs_size_kib}KiB -t 0:8304 -c 0:root atlas.img
 	sudo partprobe "$loopback"
-	sudo dd if=root.sqfs of="${loopback}p4" bs=64K status=progress
+	if [[ ! -e "${loopback}p2" ]]; then
+		echo "${loopback}p2 does not exist after partprobe." 1>&2
+		exit 1
+	fi
+	sudo dd if=root.sqfs of="${loopback}p2" bs=64K status=progress
 fi
 
 # Shrink partition table.
 backup_header_size_s=34
 sector_size=512
 next_free_block=$(sgdisk -f atlas.img)
+if [[ $next_free_block = "$backup_header_size_s" ]]; then
+	echo Next free block is primary GPT header. Has the file already been truncated? 1>&2
+	exit 1
+fi
 backup_header_start_s=$next_free_block
 backup_header_end_s=$((backup_header_start_s + backup_header_size_s))
 backup_header_end_b=$((backup_header_end_s * sector_size))
@@ -127,8 +133,18 @@ truncate -s "$backup_header_end_b" atlas.img
 sgdisk -e atlas.img
 
 # Install bootloader.
-sudo bootctl install --image=atlas.img
+sudo bootctl install --image=atlas.img || {
+	sudo mkdir root/efi
+	sudo mount "${loopback}p1" root/efi
+	sudo bootctl install --root=root --no-variables
+	sudo umount root/efi
+}
 
 mv atlas.img "$1"
 
 sudo rm -rf "$tmp"
+
+cleanup
+trap - EXIT
+
+echo Success
